Initial commit to github
Code to take MIDI files of polyphonic music and split it up into its
constituent voices, using strictly empirical methods (information on
frequencies in context of a previous note). Using Matlab and the MIDI
toolbox (included in this repository) by Petri Toiviainen and Tuomas
Eerola, Department of Music, University of Jyväskylä, Finland:
https://www.jyu.fi/hum/laitokset/musiikki/en/research/coe/materials/midi
toolbox . This repository also contains the ICMC 2008 paper on this
work and the corresponding PPT presentation.